feat: refactor customAnalyzer package for consistent naming
Refactored the customAnalyzer package and its references to use consistent snake_case naming for improved code readability and alignment with Go naming conventions.
